test(frontend): realtime python bridge e2e via launched frontend + worker
Replace the in-process realtime WebSocket e2e (which constructed an
HttpService and called the test-only start_discovery_watcher) with a
two-process test that mirrors how the system actually runs: launch the real
dynamo.frontend (which wires its own discovery watcher on startup) and a
separate Python worker that register_model(Realtime + Text) and serves the
bidirectional engine via serve_bidirectional_endpoint. The frontend
discovers the worker over etcd and installs the typed realtime PushRouter.

The worker's ManagedProcess health check gates on the model appearing in
the frontend's /v1/models, so the WebSocket session can open immediately
once the fixture yields. Uses aiohttp's WebSocket client (a core dep)
rather than websockets. Lives under tests/frontend alongside the existing
launched-frontend e2e infrastructure.
