+++ tests/frontend/realtime_echo_worker.py
@@ -1,0 +1,124 @@
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+
+"""Realtime bidirectional echo worker for the Python bridge e2e test.
+
+Registers a ``ModelType.Realtime`` + ``ModelInput.Text`` model and serves a
+Python ``async def generate(request_stream, context)`` bidirectional engine
+via ``serve_bidirectional_endpoint``. A launched ``dynamo.frontend`` discovers
+this worker over etcd and installs a typed realtime PushRouter to it.
+"""
+
+from __future__ import annotations
+
+import asyncio
+import uuid
+
+import uvloop
+
+from dynamo.llm import ModelInput, ModelType, register_model
+from dynamo.runtime import DistributedRuntime
+from tests.frontend.test_realtime_python_bridge import ENDPOINT_PATH, MODEL_NAME
+
+
+def _event_id() -> str:
+    return f"event_{uuid.uuid4().hex}"
+
+
+def _response_payload(response_id: str, status: str) -> dict:
+    """Minimal `RealtimeResponse` payload accepted by the frontend's typed reader.
+
+    The typed reader requires `id`, `max_output_tokens`, `object`, `output`,
+    `output_modalities`, and `status` to be present, so they are minted verbatim.
+    """
+    return {
+        "id": response_id,
+        "max_output_tokens": "inf",
+        "object": "realtime.response",
+        "output": [],
+        "output_modalities": ["audio"],
+        "status": status,
+    }
+
+
+async def _python_realtime_echo(request_stream, context):
+    """
+    Realtime echo semantics expressed as a Python bidirectional engine:
+
+      - `session.update` -> `session.updated` echoing the session block.
+      - `input_audio_buffer.append` -> `response.created` ->
+        `response.output_audio.delta` -> `response.output_audio.done` ->
+        `response.done`. The single delta carries the full audio payload.
+      - Anything else -> `error` event with a stable code.
+    """
+    async for client_event in request_stream:
+        if context.is_stopped():
+            return
+        etype = client_event.get("type") if isinstance(client_event, dict) else None
+
+        if etype == "session.update":
+            yield {
+                "type": "session.updated",
+                "event_id": _event_id(),
+                "session": client_event.get("session"),
+            }
+        elif etype == "input_audio_buffer.append":
+            audio = client_event.get("audio", "")
+            response_id = f"resp_{uuid.uuid4().hex}"
+            item_id = f"item_{uuid.uuid4().hex}"
+
+            yield {
+                "type": "response.created",
+                "event_id": _event_id(),
+                "response": _response_payload(response_id, "in_progress"),
+            }
+            yield {
+                "type": "response.output_audio.delta",
+                "event_id": _event_id(),
+                "response_id": response_id,
+                "item_id": item_id,
+                "output_index": 0,
+                "content_index": 0,
+                "delta": audio,
+            }
+            yield {
+                "type": "response.output_audio.done",
+                "event_id": _event_id(),
+                "response_id": response_id,
+                "item_id": item_id,
+                "output_index": 0,
+                "content_index": 0,
+            }
+            yield {
+                "type": "response.done",
+                "event_id": _event_id(),
+                "response": _response_payload(response_id, "completed"),
+            }
+        else:
+            yield {
+                "type": "error",
+                "event_id": _event_id(),
+                "error": {
+                    "type": "invalid_request_error",
+                    "code": "unsupported_client_event",
+                    "message": f"python realtime echo does not support {etype}",
+                },
+            }
+
+
+async def main() -> None:
+    """Register the realtime echo model and serve the bidirectional engine."""
+    runtime = DistributedRuntime(asyncio.get_running_loop(), "etcd", "tcp")
+    endpoint = runtime.endpoint(ENDPOINT_PATH)
+    await register_model(
+        ModelInput.Text,
+        ModelType.Realtime,
+        endpoint,
+        MODEL_NAME,
+        model_name=MODEL_NAME,
+    )
+    await endpoint.serve_bidirectional_endpoint(_python_realtime_echo)
+
+
+if __name__ == "__main__":
+    uvloop.run(main())
